+++ src/Sentry.Unity/SentryUnity.cs
@@ -1,7 +1,9 @@
 using System;
+using System.IO;
 using System.ComponentModel;
 using Sentry.Extensibility;
 using Sentry.Unity.Integrations;
+using UnityEngine;
 
 namespace Sentry.Unity
 {
@@ -10,6 +12,8 @@ namespace Sentry.Unity
     /// </summary>
     public static class SentryUnity
     {
+        private static FileStream? _lockFile;
+
         /// <summary>
         /// Initializes Sentry Unity SDK while configuring the options.
         /// </summary>
@@ -24,19 +28,53 @@ namespace Sentry.Unity
         /// <summary>
         /// Initializes Sentry Unity SDK while providing an options object.
         /// </summary>
-        /// <param name="sentryUnityOptions">The options object.</param>
+        /// <param name="options">The options object.</param>
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public static void Init(SentryUnityOptions sentryUnityOptions)
+        public static void Init(SentryUnityOptions options)
         {
-            SentryOptionsUtility.TryAttachLogger(sentryUnityOptions);
-            if (sentryUnityOptions.ShouldInitializeSdk())
+            SentryOptionsUtility.TryAttachLogger(options);
+            if (options.ShouldInitializeSdk())
             {
-                sentryUnityOptions.DiagnosticLogger?.LogDebug(sentryUnityOptions.ToString());
-                var sentryDotNet = SentrySdk.Init(sentryUnityOptions);
+                // On Standalone, we disable cache dir in case multiple app instances run over the same path.
+                // Note: we cannot use a named Mutex, because Unit doesn't support it. Instead, we create a file with `FileShare.None`.
+                // https://forum.unity.com/threads/unsupported-internal-call-for-il2cpp-mutex-createmutex_internal-named-mutexes-are-not-supported.387334/
+                if (ApplicationAdapter.Instance.Platform is RuntimePlatform.WindowsPlayer && options.CacheDirectoryPath is not null)
+                {
+                    try
+                    {
+                        _lockFile = new FileStream(Path.Combine(options.CacheDirectoryPath, "sentry-unity.lock"), FileMode.OpenOrCreate,
+                                FileAccess.ReadWrite, FileShare.None);
+                    }
+                    catch (Exception ex)
+                    {
+                        options.DiagnosticLogger?.LogWarning("An exception was thrown while trying to " +
+                            "acquire a lockfile on the config directory: .NET event cache will be disabled.", ex);
+                        options.CacheDirectoryPath = null;
+                        options.AutoSessionTracking = false;
+                    }
+                }
+
+                var sentryDotNet = SentrySdk.Init(options);
                 ApplicationAdapter.Instance.Quitting += () =>
                 {
-                    sentryUnityOptions.DiagnosticLogger?.LogDebug("Closing the sentry-dotnet SDK");
-                    sentryDotNet.Dispose();
+                    options.DiagnosticLogger?.LogDebug("Closing the sentry-dotnet SDK");
+                    try
+                    {
+                        sentryDotNet.Dispose();
+                    }
+                    finally
+                    {
+                        try
+                        {
+                            // We don't really need to close, Windows would release the lock anyway, but let's be nice.
+                            _lockFile?.Close();
+                        }
+                        catch (Exception ex)
+                        {
+                            options.DiagnosticLogger?.Log(SentryLevel.Warning,
+                                "Exception while releasing the lockfile on the config directory.", ex);
+                        }
+                    }
                 };
             }
         }